fix: incorrect decoding of query parameters
decodeURI was called too early and while it works for the host and path,
it messes up the query parameters of an url, making it impossible to
decode that part correctly.

This patch instead calls decodeURIComponent on the parts of the url
separately at the last possible time to ensure proper decoding.

The host, path, query and fragment parts are affected.